Fixed an issue with state completely overwriting on each update.
